refactor: extract config-file and template
